+++ src/main/java/no/nav/familie/ks/sak/app/behandling/regel/mvp/barnehage/BarnehageVilkår.java
@@ -1,12 +1,11 @@
-package no.nav.familie.ks.sak.app.behandling.vilkår.medlemskap;
+package no.nav.familie.ks.sak.app.behandling.regel.mvp.barnehage;
 
 import no.nav.familie.ks.sak.app.behandling.domene.kodeverk.VilkårType;
 import no.nav.familie.ks.sak.app.behandling.fastsetting.Faktagrunnlag;
 import no.nav.familie.ks.sak.app.behandling.domene.kodeverk.årsak.VilkårIkkeOppfyltÅrsak;
 import no.nav.familie.ks.sak.app.behandling.vilkår.InngangsvilkårRegel;
 import no.nav.familie.ks.sak.app.behandling.vilkår.Sluttpunkt;
-import no.nav.familie.ks.sak.app.behandling.vilkår.medlemskap.regel.HattNorskStatsborgerskapFemÅr;
-import no.nav.familie.ks.sak.app.behandling.vilkår.medlemskap.regel.HarVærtBosattFemÅrINorge;
+import no.nav.familie.ks.sak.app.behandling.regel.mvp.barnehage.regel.GårIkkeIBarnehage;
 import no.nav.fpsak.nare.Ruleset;
 import no.nav.fpsak.nare.doc.RuleDocumentation;
 import no.nav.fpsak.nare.evaluation.Evaluation;
@@ -14,12 +13,12 @@ import no.nav.fpsak.nare.specification.Specification;
 import org.springframework.stereotype.Component;
 
 @Component
-@RuleDocumentation(VilkårType.Constants.MEDLEMSKAP_KODE)
-public class MedlemskapsVilkår implements InngangsvilkårRegel<Faktagrunnlag> {
+@RuleDocumentation(VilkårType.Constants.BARNEHAGE_KODE)
+public class BarnehageVilkår implements InngangsvilkårRegel<Faktagrunnlag> {
 
     @Override
     public VilkårType getVilkårType() {
-        return VilkårType.MEDLEMSKAP;
+        return VilkårType.BARNEHAGE;
     }
 
     @Override
@@ -36,12 +35,8 @@ public class MedlemskapsVilkår implements InngangsvilkårRegel<Faktagrunnlag> {
     @SuppressWarnings("unchecked")
     public Specification<Faktagrunnlag> getSpecification() {
         final var rs = new Ruleset<Faktagrunnlag>();
-        return rs.hvisRegel(HarVærtBosattFemÅrINorge.ID, "Vurder om søker har vært bosatt i Norge siste fem år")
-                .hvis(new HarVærtBosattFemÅrINorge(),
-                        rs.hvisRegel(HattNorskStatsborgerskapFemÅr.ID, "Vurder om foreldre har vært norske statsborgere siste fem år")
-                                .hvis(new HattNorskStatsborgerskapFemÅr(), Sluttpunkt.oppfylt())
-                                .ellers(Sluttpunkt.ikkeOppfylt(VilkårIkkeOppfyltÅrsak.IKKE_NORSKE_STATSBORGERE_FEM_ÅR)))
-                .ellers(Sluttpunkt.ikkeOppfylt(VilkårIkkeOppfyltÅrsak.IKKE_BOSATT_I_NORGE_FEM_ÅR));
-
+        return rs.hvisRegel(GårIkkeIBarnehage.ID, "Vurder om barnet har barnehageplass")
+                    .hvis(new GårIkkeIBarnehage(), Sluttpunkt.oppfylt())
+                    .ellers(Sluttpunkt.ikkeOppfylt(VilkårIkkeOppfyltÅrsak.BARNEHAGEPLASS));
     }
 }
