+++ src/main/java/no/nav/familie/ks/sak/app/behandling/regel/vilkår/barn/BarneVilkår.java
@@ -1,11 +1,11 @@
-package no.nav.familie.ks.sak.app.behandling.vilkår.utland;
+package no.nav.familie.ks.sak.app.behandling.regel.vilkår.barn;
 
 import no.nav.familie.ks.sak.app.behandling.domene.kodeverk.VilkårType;
-import no.nav.familie.ks.sak.app.behandling.domene.kodeverk.årsak.VilkårIkkeOppfyltÅrsak;
 import no.nav.familie.ks.sak.app.behandling.fastsetting.Faktagrunnlag;
+import no.nav.familie.ks.sak.app.behandling.domene.kodeverk.årsak.VilkårIkkeOppfyltÅrsak;
 import no.nav.familie.ks.sak.app.behandling.vilkår.InngangsvilkårRegel;
 import no.nav.familie.ks.sak.app.behandling.vilkår.Sluttpunkt;
-import no.nav.familie.ks.sak.app.behandling.vilkår.utland.regel.IkkeOppgittTilknytningUtland;
+import no.nav.familie.ks.sak.app.behandling.regel.vilkår.barn.regel.ErNorskStatsborger;
 import no.nav.fpsak.nare.Ruleset;
 import no.nav.fpsak.nare.doc.RuleDocumentation;
 import no.nav.fpsak.nare.evaluation.Evaluation;
@@ -13,12 +13,12 @@ import no.nav.fpsak.nare.specification.Specification;
 import org.springframework.stereotype.Component;
 
 @Component
-@RuleDocumentation(VilkårType.Constants.UTLAND_KODE)
-public class UtlandVilkår implements InngangsvilkårRegel<Faktagrunnlag> {
+@RuleDocumentation(VilkårType.Constants.BARN_KODE)
+public class BarneVilkår implements InngangsvilkårRegel<Faktagrunnlag> {
 
     @Override
     public VilkårType getVilkårType() {
-        return VilkårType.UTLAND;
+        return VilkårType.BARN;
     }
 
     @Override
@@ -35,8 +35,8 @@ public class UtlandVilkår implements InngangsvilkårRegel<Faktagrunnlag> {
     @SuppressWarnings("unchecked")
     public Specification<Faktagrunnlag> getSpecification() {
         final var rs = new Ruleset<Faktagrunnlag>();
-        return rs.hvisRegel(IkkeOppgittTilknytningUtland.ID, "Vurder om søker ikke har tilknytning til utland (MVP)")
-                    .hvis(new IkkeOppgittTilknytningUtland(), Sluttpunkt.oppfylt())
-                    .ellers(Sluttpunkt.ikkeOppfylt(VilkårIkkeOppfyltÅrsak.OPPGITT_TILKNYTNING_UTLAND));
+        return rs.hvisRegel(ErNorskStatsborger.ID, "Vurder om barnet har norsk statsborgerskap")
+                    .hvis(new ErNorskStatsborger(), Sluttpunkt.oppfylt())
+                    .ellers(Sluttpunkt.ikkeOppfylt(VilkårIkkeOppfyltÅrsak.BARN_IKKE_NORSK_STATSBORGER));
     }
 }
