+++ src/main/java/no/nav/familie/ks/sak/app/behandling/regel/vilkår/bosted/BostedVilkår.java
@@ -1,12 +1,12 @@
-package no.nav.familie.ks.sak.app.behandling.vilkår.bosted;
+package no.nav.familie.ks.sak.app.behandling.regel.vilkår.bosted;
 
 import no.nav.familie.ks.sak.app.behandling.domene.kodeverk.VilkårType;
 import no.nav.familie.ks.sak.app.behandling.fastsetting.Faktagrunnlag;
 import no.nav.familie.ks.sak.app.behandling.domene.kodeverk.årsak.VilkårIkkeOppfyltÅrsak;
 import no.nav.familie.ks.sak.app.behandling.vilkår.InngangsvilkårRegel;
 import no.nav.familie.ks.sak.app.behandling.vilkår.Sluttpunkt;
-import no.nav.familie.ks.sak.app.behandling.vilkår.bosted.regel.ErBarnBosattMedForeldre;
-import no.nav.familie.ks.sak.app.behandling.vilkår.bosted.regel.HarRelasjonTilBeggeForeldre;
+import no.nav.familie.ks.sak.app.behandling.regel.vilkår.bosted.regel.ErBarnBosattMedForeldre;
+import no.nav.familie.ks.sak.app.behandling.regel.vilkår.bosted.regel.HarRelasjonTilBeggeForeldre;
 import no.nav.fpsak.nare.Ruleset;
 import no.nav.fpsak.nare.doc.RuleDocumentation;
 import no.nav.fpsak.nare.evaluation.Evaluation;
