+++ src/main/java/no/nav/familie/ks/sak/app/behandling/regel/vilkår/medlemskap/MedlemskapsVilkår.java
@@ -1,11 +1,12 @@
-package no.nav.familie.ks.sak.app.behandling.vilkår.barnehage;
+package no.nav.familie.ks.sak.app.behandling.regel.vilkår.medlemskap;
 
 import no.nav.familie.ks.sak.app.behandling.domene.kodeverk.VilkårType;
 import no.nav.familie.ks.sak.app.behandling.fastsetting.Faktagrunnlag;
 import no.nav.familie.ks.sak.app.behandling.domene.kodeverk.årsak.VilkårIkkeOppfyltÅrsak;
 import no.nav.familie.ks.sak.app.behandling.vilkår.InngangsvilkårRegel;
 import no.nav.familie.ks.sak.app.behandling.vilkår.Sluttpunkt;
-import no.nav.familie.ks.sak.app.behandling.vilkår.barnehage.regel.GårIkkeIBarnehage;
+import no.nav.familie.ks.sak.app.behandling.regel.vilkår.medlemskap.regel.HattNorskStatsborgerskapFemÅr;
+import no.nav.familie.ks.sak.app.behandling.regel.vilkår.medlemskap.regel.HarVærtBosattFemÅrINorge;
 import no.nav.fpsak.nare.Ruleset;
 import no.nav.fpsak.nare.doc.RuleDocumentation;
 import no.nav.fpsak.nare.evaluation.Evaluation;
@@ -13,12 +14,12 @@ import no.nav.fpsak.nare.specification.Specification;
 import org.springframework.stereotype.Component;
 
 @Component
-@RuleDocumentation(VilkårType.Constants.BARNEHAGE_KODE)
-public class BarnehageVilkår implements InngangsvilkårRegel<Faktagrunnlag> {
+@RuleDocumentation(VilkårType.Constants.MEDLEMSKAP_KODE)
+public class MedlemskapsVilkår implements InngangsvilkårRegel<Faktagrunnlag> {
 
     @Override
     public VilkårType getVilkårType() {
-        return VilkårType.BARNEHAGE;
+        return VilkårType.MEDLEMSKAP;
     }
 
     @Override
@@ -35,8 +36,12 @@ public class BarnehageVilkår implements InngangsvilkårRegel<Faktagrunnlag> {
     @SuppressWarnings("unchecked")
     public Specification<Faktagrunnlag> getSpecification() {
         final var rs = new Ruleset<Faktagrunnlag>();
-        return rs.hvisRegel(GårIkkeIBarnehage.ID, "Vurder om barnet har barnehageplass")
-                    .hvis(new GårIkkeIBarnehage(), Sluttpunkt.oppfylt())
-                    .ellers(Sluttpunkt.ikkeOppfylt(VilkårIkkeOppfyltÅrsak.BARNEHAGEPLASS));
+        return rs.hvisRegel(HarVærtBosattFemÅrINorge.ID, "Vurder om søker har vært bosatt i Norge siste fem år")
+                .hvis(new HarVærtBosattFemÅrINorge(),
+                        rs.hvisRegel(HattNorskStatsborgerskapFemÅr.ID, "Vurder om foreldre har vært norske statsborgere siste fem år")
+                                .hvis(new HattNorskStatsborgerskapFemÅr(), Sluttpunkt.oppfylt())
+                                .ellers(Sluttpunkt.ikkeOppfylt(VilkårIkkeOppfyltÅrsak.IKKE_NORSKE_STATSBORGERE_FEM_ÅR)))
+                .ellers(Sluttpunkt.ikkeOppfylt(VilkårIkkeOppfyltÅrsak.IKKE_BOSATT_I_NORGE_FEM_ÅR));
+
     }
 }
