+++ src/test/java/no/nav/familie/ks/sak/SaksbehandlingTest.java
@@ -5,8 +5,8 @@ import no.nav.familie.ks.sak.app.behandling.ResultatService;
 import no.nav.familie.ks.sak.app.behandling.VurderSamletTjeneste;
 import no.nav.familie.ks.sak.app.behandling.domene.kodeverk.UtfallType;
 import no.nav.familie.ks.sak.app.behandling.resultat.Vedtak;
-import no.nav.familie.ks.sak.app.behandling.vilkår.medlemskap.MedlemskapsVilkår;
-import no.nav.familie.ks.sak.app.behandling.vilkår.barn.BarneVilkår;
+import no.nav.familie.ks.sak.app.behandling.regel.vilkår.medlemskap.MedlemskapsVilkår;
+import no.nav.familie.ks.sak.app.behandling.regel.vilkår.barn.BarneVilkår;
 import no.nav.familie.ks.sak.app.grunnlag.OppslagTjeneste;
 import no.nav.familie.ks.sak.config.JacksonJsonConfig;
 import org.junit.Test;
