+++ src/test/java/no/nav/familie/ks/sak/app/behandling/PeriodeOppretterTest.java
@@ -5,10 +5,10 @@ import no.nav.familie.ks.sak.Saksbehandling;
 import no.nav.familie.ks.sak.app.behandling.fastsetting.Faktagrunnlag;
 import no.nav.familie.ks.sak.app.behandling.domene.kodeverk.UtfallType;
 import no.nav.familie.ks.sak.app.behandling.resultat.Vedtak;
-import no.nav.familie.ks.sak.app.behandling.vilkår.medlemskap.MedlemskapsVilkår;
-import no.nav.familie.ks.sak.app.behandling.vilkår.barn.BarneVilkår;
-import no.nav.familie.ks.sak.app.behandling.vilkår.barnehage.BarnehageVilkår;
-import no.nav.familie.ks.sak.app.behandling.vilkår.bosted.BostedVilkår;
+import no.nav.familie.ks.sak.app.behandling.regel.vilkår.medlemskap.MedlemskapsVilkår;
+import no.nav.familie.ks.sak.app.behandling.regel.vilkår.barn.BarneVilkår;
+import no.nav.familie.ks.sak.app.behandling.regel.mvp.barnehage.BarnehageVilkår;
+import no.nav.familie.ks.sak.app.behandling.regel.vilkår.bosted.BostedVilkår;
 import no.nav.familie.ks.sak.app.grunnlag.OppslagTjeneste;
 import no.nav.familie.ks.sak.config.JacksonJsonConfig;
 import org.junit.Test;
