+++ src/test/java/no/nav/familie/ks/sak/app/behandling/vilkår/BarnehageVilkårTest.java
@@ -2,7 +2,7 @@ package no.nav.familie.ks.sak.app.behandling.vilkår;
 
 import no.nav.familie.ks.sak.FaktagrunnlagBuilder;
 import no.nav.familie.ks.sak.app.behandling.fastsetting.Faktagrunnlag;
-import no.nav.familie.ks.sak.app.behandling.vilkår.barnehage.BarnehageVilkår;
+import no.nav.familie.ks.sak.app.behandling.regel.mvp.barnehage.BarnehageVilkår;
 import no.nav.fpsak.nare.evaluation.Resultat;
 import org.junit.Test;
