+++ src/test/java/no/nav/familie/ks/sak/app/behandling/vilkår/MedlemskapVilkårTest.java
@@ -2,7 +2,7 @@ package no.nav.familie.ks.sak.app.behandling.vilkår;
 
 import no.nav.familie.ks.sak.FaktagrunnlagBuilder;
 import no.nav.familie.ks.sak.app.behandling.fastsetting.Faktagrunnlag;
-import no.nav.familie.ks.sak.app.behandling.vilkår.medlemskap.MedlemskapsVilkår;
+import no.nav.familie.ks.sak.app.behandling.regel.vilkår.medlemskap.MedlemskapsVilkår;
 import no.nav.fpsak.nare.evaluation.Resultat;
 import org.junit.Test;
